feat: add list first team on every booth
