+++ src/lib/interfaces/participant.interface.ts
@@ -1,0 +1,5 @@
+export interface IParticipant {
+  index: number;
+  name: string;
+  type: string;
+}
